moved `prediction_intervals` module
