+++ docs/source/api_reference/experimental.rst
@@ -26,22 +26,3 @@ Classification of time-series:
    classification.PredictabilityAnalyzer
    classification.feature_extraction.TSFreshFeatureExtractor
    classification.feature_extraction.WEASELFeatureExtractor
-
-Prediction Intervals:
-
-.. autosummary::
-   :toctree: api/
-   :template: class.rst
-
-   prediction_intervals.BasePredictionIntervals
-   prediction_intervals.NaiveVariancePredictionIntervals
-   prediction_intervals.ConformalPredictionIntervals
-   prediction_intervals.EmpiricalPredictionIntervals
-
-Prediction Intervals utilities:
-
-.. autosummary::
-   :toctree: api/
-   :template: base.rst
-
-   prediction_intervals.utils.residuals_matrices
