+++ docs/source/api_reference/prediction_intervals.rst
@@ -1,0 +1,39 @@
+.. _prediction_intervals:
+
+Prediction Intervals
+====================
+
+.. automodule:: etna.prediction_intervals
+    :no-members:
+    :no-inherited-members:
+
+API details
+-----------
+
+.. currentmodule:: etna.prediction_intervals
+
+Base classes:
+
+.. autosummary::
+   :toctree: api/
+   :template: class.rst
+
+   BasePredictionIntervals
+
+Intervals classes:
+
+.. autosummary::
+   :toctree: api/
+   :template: class.rst
+
+   NaiveVariancePredictionIntervals
+   ConformalPredictionIntervals
+   EmpiricalPredictionIntervals
+
+Utilities:
+
+.. autosummary::
+   :toctree: api/
+   :template: base.rst
+
+   utils.residuals_matrices
