--- etna/experimental/prediction_intervals/__init__.py
@@ -1,4 +1,0 @@
-from etna.experimental.prediction_intervals.base import BasePredictionIntervals
-from etna.experimental.prediction_intervals.conformal import ConformalPredictionIntervals
-from etna.experimental.prediction_intervals.empirical import EmpiricalPredictionIntervals
-from etna.experimental.prediction_intervals.naive_variance import NaiveVariancePredictionIntervals
