+++ etna/prediction_intervals/__init__.py
@@ -1,0 +1,4 @@
+from etna.prediction_intervals.base import BasePredictionIntervals
+from etna.prediction_intervals.conformal import ConformalPredictionIntervals
+from etna.prediction_intervals.empirical import EmpiricalPredictionIntervals
+from etna.prediction_intervals.naive_variance import NaiveVariancePredictionIntervals
